Finished Fixes

+++ web/src/Pages/Authentication/SignUp.js
@@ -1,4 +1,7 @@
 import React, { useState } from 'react';
+import {
+    useHistory
+} from 'react-router-dom';
 import { storage } from '../../Firebase/index';
 import { Fetch } from '../../Tools';
 
@@ -12,6 +15,8 @@ const SignUp = (props) => {
     const [url, setUrl] = useState("");
     const [progress, setProgress] = useState(0);
     const [bio, setBio] = useState('');
+
+    const history = useHistory();
 
     const handleChange = e => {
         if (e.target.files[0]) {
@@ -46,20 +51,40 @@ const SignUp = (props) => {
 
     const submit = () => {
         const scoped = async () => {
-            const res = await Fetch(`
-                mutation {
-                    signup(handle: "${handle}", email: "${email}", password: "${pass}", imageUrl: "${url ? url : ""}", bio: "${bio}") {
-                        password
-                    }
+            if(url !== "") {
+                const res = await Fetch(`
+                    mutation {
+                        signup(handle: "${handle}", email: "${email}", password: "${pass}", imageUrl: "${url}", bio: "${bio}") {
+                            password
+                        }
+                    } 
+                `);
+                if(res.signup !== undefined && res.signup !== null) {
+                    sessionStorage.setItem('token', res.signup.password);
+                    history.push('/home');
+                } else {
+                    alert('Something went wrong. Please try again');
                 }
-            `)
-
-            console.log(res);
-            if(res.signup !== undefined && res.signup !== null) {
-                console.log()
-                sessionStorage.setItem('token', res.signup.password);
             } else {
-                alert('Something went wrong. Please try again');
+                const bool = window.confirm('Are you sure you don\' want a profile picture');
+                if(bool == true) {
+                    const res = await Fetch(`
+                        mutation {
+                            signup(handle: "${handle}", email: "${email}", password: "${pass}", bio: "${bio}") {
+                                password
+                            }
+
+                        }
+                    `)
+                    if(res.signup !== undefined && res.signup !== null) {
+                        sessionStorage.setItem('token', res.signup.password);
+                        history.push('/home');
+                    } else {
+                        alert('Something went wrong. Please try again');
+                    }
+                } else {
+                    window.location.reload(false);
+                }
             }
         }
 
@@ -70,13 +95,19 @@ const SignUp = (props) => {
         }
     }
 
+    const handleKeys = (e) => {
+        if(e.key == "Enter") {
+            submit();
+        }
+    }
+
     return (
         <div>
             <input value={handle} placeholder="User Handle" onChange={e => setHandle(e.target.value)} />
             <input value={email} placeholder="Email" onChange={e => setEmail(e.target.value)} />
             <input value={pass} placeholder="Password" onChange={e => setPass(e.target.value)} type="password" />
             <input value={confirm} placeholder="Confirm Password" onChange={e => setConfirm(e.target.value)} type="password" />
-            <textarea onChange={e => setBio(e.target.value)} placeholder="Biography" >
+            <textarea onChange={e => setBio(e.target.value)} onKeyDown={handleKeys} placeholder="Biography" >
             
             </textarea> 
             <div>
